zverinye_tropy_bot/requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.30
python-dotenv==1.0.1
